Add delta properties to doubles to improve test coverage to 100%

+++ src/Cup.java
@@ -20,8 +20,7 @@ public class Cup {
         this.liquidType = liquidType;
     }
 
-
-    public void setPercentFull(int percentFull) {
+    public void setPercentFull(double percentFull) {
         this.percentFull = percentFull;
     }
 }
